Move the annotation ordering endpoint test into test_adjacents.py
The test lived in tests/api/routes/api/annotations/, which mirrors a source
directory holding only create_annotation.py. It asserts on the adjacents
endpoint, so it belongs with the other get_adjacent_samples tests.
